🔧 fix: SwiftLint strict budgets — split tests + disable directives (#210)
Three violations surfaced by full-suite --strict after PR B code landed:

1. `ModelManagerTests+MultiModel.swift` grew to 481 lines. Split
   PR B additions (delete strict-guard contract, picker gate,
   cancelDownload idempotency) into a sibling
   `ModelManagerTests+StrictContracts.swift` per testing.md's
   sibling-extension pattern — no new `@Suite`, still joins the
   serialized `ModelManagerTests` suite.

2. `PasturaApp.swift` now 404 lines after `.needsModelSelection`
   plumbing. File-level `swiftlint:disable file_length` with a
   rationale comment (consistent with `ModelManager.swift`). Splitting
   would require exporting file-private enums (`AppState`,
   `AppStateKind`, `DeepLinkBlockReason`) across files for a
   mechanical line-count win that weakens encapsulation.

3. `ModelManager`'s class body grew to 259 lines after the strict
   `deleteModel` + `shouldShowInitialModelPicker` additions. Added
   `swiftlint:disable:this type_body_length` on the declaration;
   the existing file-level rationale already covers why.

No behavior changes. 17 tests still pass in the strict-contract split.
